requirements.txt
ipykernel
openai
weasyprint
jinja2
PyYAML
python-dotenv
fpdf
streamlit